requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
Jinja2==2.7.1
MarkupSafe==0.18
SQLAlchemy==1.1.14
Tempita==0.5.2
WTForms==2.1
Werkzeug==0.9.4
click==6.7
decorator==4.1.2
dominate==2.3.1
itsdangerous==0.23
pbr==3.1.1
pywebview==1.7
six==1.10.0
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
visitor==0.1.3
